Admin APi is almost done

+++ src/main/java/com/imdad/binding/DashboardCards.java
@@ -10,7 +10,7 @@ public class DashboardCards {
     private Long totalPlans;
     private Long approvedCitizens;
     private Long deniedCitizens;
-    private Long totalBenifitGiven;
+    private Double totalBenifitGiven;
